requirements.txt
matplotlib
pandas~=2.2.1
plotly~=5.19.0
pytest
python-multipart
python-dotenv
freezegun
psycopg2-binary
unittest2
uvicorn
argparse~=1.4.0
numpy~=1.26.4
seaborn
holidays
tqdm
requests
scikit-learn
argparse
tabulate